(#178) refactor: Router 관련 구조 변경
What
1. URLSession 을 사용하는 Router 와 단순히 Json 데이터만 반환하는 Mock Router 를 위해 이 둘의 공통적인 속성을 정의하는 Routable 프로토콜을 만들었습니다
2. 클래스에서 제네릭을 사용했던 기존 구조는 프로토콜을 통해 묶는게 불가능하기 때문에 Router 클래스의 제네릭을 없앴습니다
