+++ iOS/IssueTracker/IssueTracker/Model/Network/Service/Router.swift
@@ -7,48 +7,15 @@
 
 import UIKit
 
-
-enum APIError: Error {
-    case data,
-         decodingJSON
-    case redirection,  // 300 번
-         client,        // 400 번
-         server,        // 500 번
-         failed
-}
-
-final class Router<EndPoint: EndPointable> {
+final class Router: Routable {
     private var task: URLSessionTask?
     
-    func openSite(from route: EndPoint) {
+    func openSite(from route: EndPointable) {
         guard let url = URL(string: "\(route.baseURL)\(route.query)") else { return }
         UIApplication.shared.open(url)
     }
     
-    func handleNetworkResponseError(_ response: HTTPURLResponse) -> APIError? {
-        switch response.statusCode {
-        case 200...299: return nil
-        case 300...399: return .redirection
-        case 401...500: return .client
-        case 501...599: return .server
-        default: return .failed
-        }
-    }
-    
-    private func configureRequest(from route: EndPoint) -> URLRequest {
-        let url = route.query == "" ? route.baseURL : route.baseURL.appendingPathComponent(route.query)
-        var request = URLRequest(url: url)
-    
-        request.httpMethod = route.httpMethod?.rawValue
-        request.httpBody = route.bodies?.encode().data(using: String.Encoding.utf8)
-        route.headers?.forEach { key, value in
-            request.setValue(value, forHTTPHeaderField: key)
-        }
-        
-        return request
-    }
-   
-    func request<T: Decodable>(route: EndPoint, completionHandler: ((Result<T,APIError>) -> Void)? = nil) {
+    func request<T: Decodable>(route: EndPointable, completionHandler: ((Result<T,APIError>) -> Void)? = nil) {
         let session = URLSession.shared
         
         let request = configureRequest(from: route)
@@ -92,4 +59,6 @@ extension Dictionary {
         }.joined(separator: "&")
     }
 }
+
+
